+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.1.7-pre
-*
+* Fix http#req delete method handling
 
 ### 0.1.6
 * Add http#req no proxy hosts handling, localhost and 127.0.0 will ignore proxy by default